// src/types/package.ts
export interface SessionPackage {
  id: string;          // PKG01, PKG02...
  name: string;        // Gói 10 buổi, Gói 20 buổi...
  sessionCount: number;// 10, 20, 30...
  price: number;       // 1000000, 1800000...
  isActive: boolean;   // true / false
  description?: string;
  createdAt?: string;
  updatedAt?: string;
}